feat: use consolidated backend function
